Node check on startup, fixed deploy issue, error_msg issue in shell_command, redirecting err to std out in bash scripts
Nodes are pinged before any setup occurs on that node. If there is only 1 runner node configured this will cause the process to hang currently.
Removed log4j and missed the deploy part in the Rakefile. This is fixed.
If there was a non 0 exit code on a bash script the error formatter was improperly handling the format.Fixed this so it looks good.
Error out on shell_commands is now redirected to stdout automatically. You will see this in the log file now.