Formatting errors better, Updated documentation on setup files, Fixed bad bug with running the Erlang VM as a non root user.
The issue with running the Erlang VM detached is that if you start it in an ssh session on OSX it will loose the launchctl manager when you exit ssh.
This will cause the VM to no longer be able to use DirectoryServices etc and that means no SSH for git.
The way it's started now the manager is Background and it appears to be working good in 1 day of testing.  The way the vm's are being launched will probably change again in the future.

Setting default SSH user to the currently logged in user.This is still overriden per node in node_config.txt if needed
Test files to run are now sorted first by the time taken from the last run, next on if they are using shared specs, and finally on size. This will be cleaned up in future as this much logic is no longer needed with ordering off previous run.
Now using spawn_link between the local runner_node_prep process and the remote one.
Removed throws in shell_command as exit is more Erlang like.